backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.8,<3.0
python-dotenv>=1.0.0,<2.0.0
sentence-transformers>=3.0.0,<4.0.0
numpy>=1.26.0,<3.0.0